+++ packages/gatsby-admin/src/components/providers.tsx
@@ -2,19 +2,87 @@ import React from "react"
 import { Provider, Client } from "urql"
 import { ThemeProvider, getTheme } from "gatsby-interface"
 import { ThemeProvider as StrictUIProvider } from "strict-ui"
-import { Spinner } from "theme-ui"
+import { Spinner, merge } from "theme-ui"
 import { createUrqlClient } from "../urql-client"
+import "normalize.css"
 
 const baseTheme = getTheme()
 
-const theme = {
-  ...baseTheme,
+const recipesTheme = merge(baseTheme, {
   colors: {
-    ...baseTheme.colors,
+    background: `white`,
+  },
+  styles: {
+    h1: {
+      fontSize: 6,
+      fontFamily: `heading`,
+      fontWeight: `heading`,
+      mt: 0,
+      mb: 4,
+    },
+    h2: {
+      fontSize: 5,
+      fontFamily: `heading`,
+      fontWeight: `heading`,
+      mt: 0,
+      mb: 4,
+    },
+    p: {
+      color: baseTheme.tones.NEUTRAL.dark,
+      fontSize: 2,
+      fontFamily: `body`,
+      fontWeight: `body`,
+      mt: 0,
+      mb: 4,
+      lineHeight: 1.45,
+    },
+    pre: {
+      fontFamily: baseTheme.fonts.monospace,
+      fontSize: 0,
+      lineHeight: 1.45,
+      mt: 0,
+      mb: 6,
+      whiteSpace: `pre-wrap`,
+    },
+    inlineCode: {
+      backgroundColor: `hsla(0,0%,0%,0.06)`,
+      color: baseTheme.tones.NEUTRAL.darker,
+      borderRadius: `3px`,
+      py: `0.2em`,
+      px: `0.2em`,
+      fontSize: `90%`,
+    },
+    ol: {
+      color: baseTheme.tones.NEUTRAL.dark,
+      paddingLeft: 8,
+      mt: 0,
+      mb: 6,
+      fontFamily: `body`,
+      fontWeight: `body`,
+    },
+    ul: {
+      color: baseTheme.tones.NEUTRAL.dark,
+      paddingLeft: 8,
+      mt: 0,
+      mb: 6,
+      fontFamily: `body`,
+      fontWeight: `body`,
+    },
+    li: {
+      color: baseTheme.tones.NEUTRAL.dark,
+      mb: 2,
+      fontFamily: `body`,
+      fontWeight: `body`,
+      lineHeight: 1.6,
+    },
+  },
+})
+
+const theme = merge(recipesTheme, {
+  colors: {
     background: baseTheme.colors.primaryBackground,
   },
   fonts: {
-    ...baseTheme.fonts,
     // We want to use inter for all text on the page, no more futura!
     brand: baseTheme.fonts.sans,
     heading: baseTheme.fonts.sans,
@@ -61,7 +129,7 @@ const theme = {
       backgroundColor: `grey.5`,
     },
   },
-}
+})
 
 const GraphQLProvider: React.FC<{}> = ({ children }) => {
   const [status, setStatus] = React.useState(`loading`)