+++ packages/gatsby-core-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.3.17...gatsby-core-utils@1.3.18) (2020-08-28)
+
+**Note:** Version bump only for package gatsby-core-utils
+
 ## [1.3.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.3.16...gatsby-core-utils@1.3.17) (2020-08-26)
 
 **Note:** Version bump only for package gatsby-core-utils