+++ packages/gatsby-plugin-sharp/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.30](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.6.29...gatsby-plugin-sharp@2.6.30) (2020-08-28)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
 ## [2.6.29](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.6.28...gatsby-plugin-sharp@2.6.29) (2020-08-26)
 
 **Note:** Version bump only for package gatsby-plugin-sharp